+++ pax/data_model.py
@@ -2,8 +2,9 @@
 Python skeleton for data structure
 Extends python object to do a few tricks
 """
-import numpy as np
 import json
+
+import numpy as np
 
 from pax.utils import Memoize
 
@@ -49,12 +50,14 @@ class Model(object):
 
     def get_fields_data(self):
         """Iterator over (key, value) tuples of all user-specified fields
+        Returns keys in lexical order
         """
+        # TODO: increase performance by pre-sorting keys?
         # self.__dict__.items() does not return default values set in class declaration
         # Hence we need something more complicated
         class_dict = self.__class__.__dict__
         self_dict = self.__dict__
-        for field_name in class_dict.keys():
+        for field_name in sorted(class_dict.keys()):
             if field_name in self_dict:
                 # The instance has a value for this field: return it
                 yield (field_name, self_dict[field_name])
@@ -65,8 +68,8 @@ class Model(object):
                 value_in_class = class_dict[field_name]
                 if callable(value_in_class):
                     continue    # No, is a method
-                if isinstance(value_in_class, property):
-                    continue    # No, is a property
+                if isinstance(value_in_class, (property, classmethod)):
+                    continue    # No, is a property or classmethod
                 # Yes, yield the class-level value
                 yield (field_name, value_in_class)