+++ pax/exceptions.py
@@ -1,3 +1,7 @@
 
 class OccurrenceBeyondEventError(Exception):
     pass
+
+
+class OutputFileAlreadyExistsError(Exception):
+    pass