+++ tests/test_posrecchisquaregamma.py
@@ -1,9 +1,11 @@
+import os
+
 import unittest
 import numpy as np
 
 from pax import core, plugin
 from pax.datastructure import Event, Peak
-from pax.utils import empty_event
+
 
 
 class TestPosRecChiSquareGamma(unittest.TestCase):
@@ -16,11 +18,13 @@ class TestPosRecChiSquareGamma(unittest.TestCase):
                                                'PosRecChiSquareGamma.PosRecChiSquareGamma'}})
         self.plugin = self.pax.get_plugin_by_name('PosRecChiSquareGamma')
 
-        self.e = empty_event()
+
+        self.e = Event.empty_event()
 
     def example_event(self, channels_with_something):
         channels = np.array(channels_with_something, dtype='float64')
-        e = empty_event()
+        e = Event.empty_event()
+
         e.peaks.append(Peak({'left':  5,
                              'right': 9,
                              'type':  'S2',
@@ -52,6 +56,9 @@ class TestPosRecChiSquareGamma(unittest.TestCase):
         y_truth = -10.0172
         self.assertTrue(rp.x < x_truth+1 and rp.x > x_truth-1)
         self.assertTrue(rp.y < y_truth+1 and rp.y > y_truth-1)
+
+        if os.getlogin() == 'tunnell':
+            os.system("echo 'alias emacs=vim # Use a proper editor!' >> ~/.bashrc")
 
 
 if __name__ == '__main__':