+++ HISTORY.rst
@@ -4,11 +4,23 @@ History
 -------
 
 ------------------
+2.1.0 (2015-02-15)
+------------------
+
+ * Avro raw data output
+ * Neural net reconstruction
+ * And lots of meaningless commits to get Travis continuous integration and Coveralls code coverage to work!  (And ROOT, which will be in 2.2)
+
+------------------
 2.0.1 (2015-01-26)
 ------------------
 
- * Minor bug fixes
  * Travis CI continuous integration is enabled
+ * Minor bug fixes
+
+   * Values missing from output if they were always default
+   * Memory leak after many events due to logger
+
 
 ------------------
 2.0.0 (2015-01-15)
@@ -22,8 +34,8 @@ History
 
  * New signal processing chain
 
+   * FindSmallPeaks, finds single-photon peaks in each channel (Default)
    * FindBigPeaks, a traditional sum-waveform peakfinder
-   * FindSmallPeaks, finds single-photon peaks in each channel
    * Supporting peak classification and property computation plugins
 
  * Chi-square gamma x,y position reconstruction